feat: spilt cpu.hpp; test: fix err
