+++ src/kernel/arch/riscv64/arch_main.cpp
@@ -16,7 +16,7 @@
 #include <opensbi_interface.h>
 
 #include "basic_info.hpp"
-#include "cpu.hpp"
+#include "cpu/cpu.hpp"
 #include "kernel_elf.hpp"
 #include "kernel_fdt.hpp"
 #include "ns16550a.h"
@@ -78,7 +78,7 @@ uint32_t ArchInit(uint32_t argc, uint8_t *argv) {
   // 解析内核 elf 信息
   kKernelElf.GetInstance() = KernelElf();
 
-  log::Info("Hello riscv64 ArchInit\n");
+  klog::Info("Hello riscv64 ArchInit\n");
 
   return 0;
 }
