+++ src/kernel/arch/x86_64/backtrace.cpp
@@ -14,9 +14,9 @@
  */
 
 #include "arch.h"
-#include "cpu.hpp"
-#include "sk_cstdio"
+#include "cpu/cpu.hpp"
 #include "kernel_elf.hpp"
+#include "sk_cstdio"
 #include "sk_libc.h"
 
 int backtrace(void **buffer, int size) {
