+++ test/unit_test/sk_cstdio
@@ -1,0 +1,22 @@
+
+/**
+ * @file sk_cstdio
+ * @brief 测试使用的 sk_cstdio
+ * @author Zone.N (Zone.Niuzh@hotmail.com)
+ * @version 1.0
+ * @date 2023-05-08
+ * @copyright MIT LICENSE
+ * https://github.com/Simple-XX/SimpleKernel
+ * @par change log:
+ * <table>
+ * <tr><th>Date<th>Author<th>Description
+ * <tr><td>2023-05-08<td>Zone.N<td>创建文件
+ * </table>
+ */
+
+#ifndef SIMPLEKERNEL_TEST_UNIT_TEST_SK_CSTDIO_
+#define SIMPLEKERNEL_TEST_UNIT_TEST_SK_CSTDIO_
+
+#include "cstdio"
+
+#endif /* SIMPLEKERNEL_TEST_UNIT_TEST_SK_CSTDIO_ */
